// Source/DocuEye.Web/client-app/app/api/docueye-api/models/issue.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DocuEye.Web
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



// May contain unused imports in some cases
// @ts-ignore
import type { IssueElement } from './issue-element';
// May contain unused imports in some cases
// @ts-ignore
import type { IssueRelationship } from './issue-relationship';
// May contain unused imports in some cases
// @ts-ignore
import type { IssueRule } from './issue-rule';

export interface Issue {
    'id'?: string | null;
    'workspaceId'?: string | null;
    'severity'?: number;
    'message'?: string | null;
    'rule'?: IssueRule;
    'relationship'?: IssueRelationship;
    'element'?: IssueElement;
}